fix(*): fix lack of reactivity on several components
Some components were accessing refs from repositories *directly on
setup*. This was fine when `SadModal` and `SadDrawer` used v-if to be
shown/hidden, but represents a problem now that they are always rendered
and have their visibility changed using `show` prop. To avoid issues,
this commit adds `computed` to these instances so they react to changes
on the aforementioned refs.

Also, there is an issue with forms not being cleaned or refilled. This
happes after a modal/drawer becomes invisible, or after a prop that this
form depends on changes. To address that, this commit introduces
watchers that react to changes on these dependencies.